+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added `subTitle` to the `DataGridCell` component to display additional subtitle text in the row data.
 - Added jest-reporter-log-validator to report on max allowed log counts from component rendering during pull request test check
 - Added `ActionButton` component designed to perform a specific action or task when clicked or tapped.
+- Added `neutral` size control in the `Button` component.
 
 ### Fixed
 - Fixed accessibility issue with `DataGrid` component